+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog].
 
 ### Fixed
 
+- Запуск Puppeteer от root в Docker (--no-sandbox)
+
 ### Security
 
 ### Internal